Move ColorEditorListBox and ColorEditorTabControl to their own files
